API

Documentation for SmoothInterpolation.jl's public interface.

SmoothInterpolation.SmoothedLinearInterpolation — Type

SmoothedLinearInterpolation(u, t; λ = 0.25, extrapolate = false)

The method of interpolating between the data points using a linear polynomial, with an anterval around the corner points being replaced by a smooth spline section.

Arguments

u: data points.

t: time points.

Keyword Arguments

extrapolate: boolean value to allow extrapolation. Defaults to false.

λ: The relative size of the spline interval. The interval extents a fraction λ/2 towards the neighbouring time points.

DataInterpolations.LinearInterpolation — Method

LinearInterpolation(A::SmoothedLinearInterpolation; n_samples = 10)

Converting a SmoothedLinearInterpolation object into LinearInterpolation object by sampling the spline sections. The main usage of this is that a LinearInterpolation and especially its integration inverse are much cheaper to evaluate than the original smoothed equivalents.

Arguments

A: The SmoothedLinearInterpolation object

Keyword Arguments

n_samples: The number of samples per spline section

SmoothInterpolation.invert_integral — Method

Invert the integral of a LinearInterpolation object, which yields 
LinearInterpolationIntInv object.

SmoothInterpolation.invert_integral — Method

Invert the integral of a SmoothedLinearInterpolation object, which yields 
SmoothedLinearInterpolationIntInv object.